Move to src-based layout
This moves away from a single-file/module package with the main purpose of
the ability to add a py.typed marker in the package [PEP 561].

As per discussion in #110, this also converts the 'crc' module into a
private module with only specific public entities being 'exported' in the
package's __init__.py. This change should not affect typical use case of
dependent projects where the crc package is installed and used confirming
the documentation. For running the tests it may now be required to install
the package in the virtualenv using `poetry install`.

References: #110

[PEP 561]: https://peps.python.org/pep-0561/#packaging-type-information